2024_06_24/requirements.txt
psycopg2-binary
requests
pydantic
